Trait substrate_client::backend::BlockImportOperation

pub trait BlockImportOperation<Block, H> where
    Block: BlockT,
    H: Hasher<Out = Block::Hash>, {
    type State: StateBackend<H>;
    fn state(&self) -> Result<Option<&Self::State>>;

    fn set_block_data(
        &mut self, 
        header: Block::Header, 
        body: Option<Vec<Block::Extrinsic>>, 
        justification: Option<Justification>, 
        state: NewBlockState
    ) -> Result<()>;

    fn update_cache(&mut self, cache: HashMap<Id, Vec<u8>>);

    fn update_db_storage(
        &mut self, 
        update: <Self::State as StateBackend<H>>::Transaction
    ) -> Result<()>;

    fn reset_storage(
        &mut self, 
        top: StorageOverlay, 
        children: ChildrenStorageOverlay
    ) -> Result<H::Out>;

    fn update_storage(
        &mut self, 
        update: StorageCollection, 
        child_update: ChildStorageCollection
    ) -> Result<()>;

    fn update_changes_trie(&mut self, update: MemoryDB<H>) -> Result<()>;

    fn insert_aux<I>(&mut self, ops: I) -> Result<()>
    where
        I: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>;

    fn mark_finalized(
        &mut self, 
        id: BlockId<Block>, 
        justification: Option<Justification>
    ) -> Result<()>;

    fn mark_head(&mut self, id: BlockId<Block>) -> Result<()>;
}

Block insertion operation. Keeps hold if the inserted block state and data.

Associated Types

type State: StateBackend<H>

Associated state backend type.

Loading content...
Required methods

fn state(&self) -> Result<Option<&Self::State>>

Returns pending state. Returns None for backends with locally-unavailable state data.

fn set_block_data(
&mut self,
header: Block::Header,
body: Option<Vec<Block::Extrinsic>>,
justification: Option<Justification>,
state: NewBlockState
) -> Result<()>

Append block data to the transaction.

fn update_cache(&mut self, cache: HashMap<Id, Vec<u8>>)

Update cached data.

fn update_db_storage(
&mut self,
update: <Self::State as StateBackend<H>>::Transaction
) -> Result<()>

Inject storage data into the database.

fn reset_storage(
&mut self,
top: StorageOverlay,
children: ChildrenStorageOverlay
) -> Result<H::Out>

Inject storage data into the database replacing any existing data.

fn update_storage(
&mut self,
update: StorageCollection,
child_update: ChildStorageCollection
) -> Result<()>

Set storage changes.

fn update_changes_trie(&mut self, update: MemoryDB<H>) -> Result<()>

Inject changes trie data into the database.

fn insert_aux<I>(&mut self, ops: I) -> Result<()> where
I: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>,

Insert auxiliary keys. Values are None if should be deleted.

fn mark_finalized(
&mut self,
id: BlockId<Block>,
justification: Option<Justification>
) -> Result<()>

Mark a block as finalized.

fn mark_head(&mut self, id: BlockId<Block>) -> Result<()>

Mark a block as new head. If both block import and set head are specified, set head overrides block import's best block rule.

Loading content...