Trait substrate_client::blockchain::Backend

pub trait Backend<Block: BlockT>: HeaderBackend<Block> {
    fn body(
        &self, 
        id: BlockId<Block>
    ) -> Result<Option<Vec<<Block as BlockT>::Extrinsic>>>;

    fn justification(&self, id: BlockId<Block>) -> Result<Option<Justification>>;

    fn last_finalized(&self) -> Result<Block::Hash>;

    fn cache(&self) -> Option<Arc<dyn Cache<Block>>>;

    fn leaves(&self) -> Result<Vec<Block::Hash>>;

    fn children(&self, parent_hash: Block::Hash) -> Result<Vec<Block::Hash>>;
}

Blockchain database backend. Does not perform any validation.

Required methods

fn body(
&self,
id: BlockId<Block>
) -> Result<Option<Vec<<Block as BlockT>::Extrinsic>>>

Get block body. Returns None if block is not found.

fn justification(&self, id: BlockId<Block>) -> Result<Option<Justification>>

Get block justification. Returns None if justification does not exist.

fn last_finalized(&self) -> Result<Block::Hash>

Get last finalized block hash.

fn cache(&self) -> Option<Arc<dyn Cache<Block>>>

Returns data cache reference, if it is enabled on this backend.

fn leaves(&self) -> Result<Vec<Block::Hash>>

Returns hashes of all blocks that are leaves of the block tree. in other words, that have no children, are chain heads. Results must be ordered best (longest, heighest) chain first.

fn children(&self, parent_hash: Block::Hash) -> Result<Vec<Block::Hash>>

Return hashes of all blocks that are children of the block with parent_hash.

Loading content...