Trait substrate_client::light::fetcher::Fetcher

pub trait Fetcher<Block: BlockT>: Send + Sync {
    type RemoteHeaderResult: IntoFuture<Item = Block::Header, Error = ClientError>;
    type RemoteReadResult: IntoFuture<Item = Option<Vec<u8>>, Error = ClientError>;
    type RemoteCallResult: IntoFuture<Item = Vec<u8>, Error = ClientError>;
    type RemoteChangesResult: IntoFuture<Item = Vec<(NumberFor<Block>, u32)>, Error = ClientError>;
    type RemoteBodyResult: IntoFuture<Item = Vec<Block::Extrinsic>, Error = ClientError>;
    fn remote_header(
        &self, 
        request: RemoteHeaderRequest<Block::Header>
    ) -> Self::RemoteHeaderResult;

    fn remote_read(
        &self, 
        request: RemoteReadRequest<Block::Header>
    ) -> Self::RemoteReadResult;

    fn remote_read_child(
        &self, 
        request: RemoteReadChildRequest<Block::Header>
    ) -> Self::RemoteReadResult;

    fn remote_call(
        &self, 
        request: RemoteCallRequest<Block::Header>
    ) -> Self::RemoteCallResult;

    fn remote_changes(
        &self, 
        request: RemoteChangesRequest<Block::Header>
    ) -> Self::RemoteChangesResult;

    fn remote_body(
        &self, 
        request: RemoteBodyRequest<Block::Header>
    ) -> Self::RemoteBodyResult;
}

Light client data fetcher. Implementations of this trait must check if remote data is correct (see FetchedDataChecker) and return already checked data.

Associated Types

type RemoteHeaderResult: IntoFuture<Item = Block::Header, Error = ClientError>

Remote header future.

type RemoteReadResult: IntoFuture<Item = Option<Vec<u8>>, Error = ClientError>

Remote storage read future.

type RemoteCallResult: IntoFuture<Item = Vec<u8>, Error = ClientError>

Remote call result future.

type RemoteChangesResult: IntoFuture<Item = Vec<(NumberFor<Block>, u32)>, Error = ClientError>

Remote changes result future.

type RemoteBodyResult: IntoFuture<Item = Vec<Block::Extrinsic>, Error = ClientError>

Remote block body result future.

Loading content...
Required methods

fn remote_header(
&self,
request: RemoteHeaderRequest<Block::Header>
) -> Self::RemoteHeaderResult

Fetch remote header.

fn remote_read(
&self,
request: RemoteReadRequest<Block::Header>
) -> Self::RemoteReadResult

Fetch remote storage value.

fn remote_read_child(
&self,
request: RemoteReadChildRequest<Block::Header>
) -> Self::RemoteReadResult

Fetch remote storage child value.

fn remote_call(
&self,
request: RemoteCallRequest<Block::Header>
) -> Self::RemoteCallResult

Fetch remote call result.

fn remote_changes(
&self,
request: RemoteChangesRequest<Block::Header>
) -> Self::RemoteChangesResult

Fetch remote changes ((block number, extrinsic index)) where given key has been changed at a given blocks range.

fn remote_body(
&self,
request: RemoteBodyRequest<Block::Header>
) -> Self::RemoteBodyResult

Fetch remote block body

Loading content...