Module substrate_client::runtime_api

All the functionality required for declaring and implementing runtime apis.

Enums

InitializeBlock
Before calling any runtime api function, the runtime need to be initialized at the requested block. However, some functions like execute_block or initialize_block itself don't require to have the runtime initialized at the requested block.

Traits

ApiExt
An extension for the RuntimeApi.

CallRuntimeAt
Something that can call into the runtime at a given block.

ConstructRuntimeApi
Something that can be constructed to a runtime api.

Core
The Core api trait that is mandatory for each runtime.

Metadata
The Metadata api trait that returns metadata for the runtime.

TaggedTransactionQueue
The TaggedTransactionQueue api trait for interfering with the new transaction queue.

Type Definitions

ProofRecorder
A type that records all accessed trie nodes and generates a proof out of it.