Trait substrate_client::runtime_api::ApiExt

pub trait ApiExt<Block: BlockT> {
    fn map_api_result<F: FnOnce(&Self) -> Result<R, E>, R, E>(
        &self, 
        map_call: F
    ) -> Result<R, E>
    where
        Self: Sized;

    fn runtime_version_at(&self, at: &BlockId<Block>) -> Result<RuntimeVersion>;

    fn record_proof(&mut self);

    fn extract_proof(&mut self) -> Option<Vec<Vec<u8>>>;

    fn has_api<A: RuntimeApiInfo + ?Sized>(
        &self, 
        at: &BlockId<Block>
    ) -> Result<bool>
    where
        Self: Sized,
    { ... }

    fn has_api_with<A: RuntimeApiInfo + ?Sized, P: Fn(u32) -> bool>(
        &self, 
        at: &BlockId<Block>, 
        pred: P
    ) -> Result<bool>
    where
        Self: Sized,
    { ... }
}

An extension for the RuntimeApi.

Required methods

fn map_api_result<F: FnOnce(&Self) -> Result<R, E>, R, E>(
&self,
map_call: F
) -> Result<R, E> where
Self: Sized,

The given closure will be called with api instance. Inside the closure any api call is allowed. After doing the api call, the closure is allowed to map the Result to a different Result type. This can be important, as the internal data structure that keeps track of modifications to the storage, discards changes when the Result is an Err. On Ok, the structure commits the changes to an internal buffer.

fn runtime_version_at(&self, at: &BlockId<Block>) -> Result<RuntimeVersion>

Returns the runtime version at the given block id.

fn record_proof(&mut self)

Start recording all accessed trie nodes for generating proofs.

fn extract_proof(&mut self) -> Option<Vec<Vec<u8>>>

Extract the recorded proof. This stops the proof recording.

Loading content...
Provided methods

fn has_api<A: RuntimeApiInfo + ?Sized>(
&self,
at: &BlockId<Block>
) -> Result<bool> where
Self: Sized,

Checks if the given api is implemented and versions match.

fn has_api_with<A: RuntimeApiInfo + ?Sized, P: Fn(u32) -> bool>(
&self,
at: &BlockId<Block>,
pred: P
) -> Result<bool> where
Self: Sized,

Check if the given api is implemented and the version passes a predicate.

Loading content...