Trait substrate_client::runtime_api::CallRuntimeAt

pub trait CallRuntimeAt<Block: BlockT> {
    fn call_api_at<'a, R: Encode + Decode + PartialEq, NC: FnOnce() -> Result<R, &'static str> + UnwindSafe, C: Core<Block>>(
        &self, 
        core_api: &C, 
        at: &BlockId<Block>, 
        function: &'static str, 
        args: Vec<u8>, 
        changes: &RefCell<OverlayedChanges>, 
        initialize_block: InitializeBlock<'a, Block>, 
        native_call: Option<NC>, 
        context: ExecutionContext, 
        recorder: &Option<Rc<RefCell<ProofRecorder<Block>>>>
    ) -> Result<NativeOrEncoded<R>>;

    fn runtime_version_at(&self, at: &BlockId<Block>) -> Result<RuntimeVersion>;
}

Something that can call into the runtime at a given block.

Required methods

fn call_api_at<'a, R: Encode + Decode + PartialEq, NC: FnOnce() -> Result<R, &'static str> + UnwindSafe, C: Core<Block>>(
&self,
core_api: &C,
at: &BlockId<Block>,
function: &'static str,
args: Vec<u8>,
changes: &RefCell<OverlayedChanges>,
initialize_block: InitializeBlock<'a, Block>,
native_call: Option<NC>,
context: ExecutionContext,
recorder: &Option<Rc<RefCell<ProofRecorder<Block>>>>
) -> Result<NativeOrEncoded<R>>

Calls the given api function with the given encoded arguments at the given block and returns the encoded result.

fn runtime_version_at(&self, at: &BlockId<Block>) -> Result<RuntimeVersion>

Returns the runtime version at the given block.

Loading content...