Trait substrate_client::CallExecutor

pub trait CallExecutor<B, H> where
    B: BlockT,
    H: Hasher<Out = B::Hash>,
    H::Out: Ord, {
    type Error: Error;
    fn call<O: Externalities>(
        &self, 
        id: &BlockId<B>, 
        method: &str, 
        call_data: &[u8], 
        strategy: ExecutionStrategy, 
        side_effects_handler: Option<&mut O>
    ) -> Result<Vec<u8>, Error>;

    fn contextual_call<'a, O: Externalities, IB: Fn() -> Result<()>, EM: Fn(Result<NativeOrEncoded<R>, Self::Error>, Result<NativeOrEncoded<R>, Self::Error>) -> Result<NativeOrEncoded<R>, Self::Error>, R: Encode + Decode + PartialEq, NC: FnOnce() -> Result<R, &'static str> + UnwindSafe>(
        &self, 
        initialize_block_fn: IB, 
        at: &BlockId<B>, 
        method: &str, 
        call_data: &[u8], 
        changes: &RefCell<OverlayedChanges>, 
        initialize_block: InitializeBlock<'a, B>, 
        execution_manager: ExecutionManager<EM>, 
        native_call: Option<NC>, 
        side_effects_handler: Option<&mut O>, 
        proof_recorder: &Option<Rc<RefCell<ProofRecorder<B>>>>
    ) -> Result<NativeOrEncoded<R>>
    where
        ExecutionManager<EM>: Clone;

    fn runtime_version(&self, id: &BlockId<B>) -> Result<RuntimeVersion, Error>;

    fn call_at_state<O: Externalities, S: Backend<H>, F: FnOnce(Result<NativeOrEncoded<R>, Self::Error>, Result<NativeOrEncoded<R>, Self::Error>) -> Result<NativeOrEncoded<R>, Self::Error>, R: Encode + Decode + PartialEq, NC: FnOnce() -> Result<R, &'static str> + UnwindSafe>(
        &self, 
        state: &S, 
        overlay: &mut OverlayedChanges, 
        method: &str, 
        call_data: &[u8], 
        manager: ExecutionManager<F>, 
        native_call: Option<NC>, 
        side_effects_handler: Option<&mut O>
    ) -> Result<(NativeOrEncoded<R>, S::Transaction, Option<MemoryDB<H>>), Error>;

    fn prove_at_trie_state<S: TrieBackendStorage<H>>(
        &self, 
        trie_state: &TrieBackend<S, H>, 
        overlay: &mut OverlayedChanges, 
        method: &str, 
        call_data: &[u8]
    ) -> Result<(Vec<u8>, Vec<Vec<u8>>), Error>;

    fn native_runtime_version(&self) -> Option<&NativeVersion>;

    fn prove_at_state<S: Backend<H>>(
        &self, 
        state: S, 
        overlay: &mut OverlayedChanges, 
        method: &str, 
        call_data: &[u8]
    ) -> Result<(Vec<u8>, Vec<Vec<u8>>), Error> { ... }
}

Method call executor.

Associated Types

type Error: Error

Externalities error type.

Loading content...
Required methods

fn call<O: Externalities>(
&self,
id: &BlockId<B>,
method: &str,
call_data: &[u8],
strategy: ExecutionStrategy,
side_effects_handler: Option<&mut O>
) -> Result<Vec<u8>, Error>

Execute a call to a contract on top of state in a block of given hash.

No changes are made.

fn contextual_call<'a, O: Externalities, IB: Fn() -> Result<()>, EM: Fn(Result<NativeOrEncoded<R>, Self::Error>, Result<NativeOrEncoded<R>, Self::Error>) -> Result<NativeOrEncoded<R>, Self::Error>, R: Encode + Decode + PartialEq, NC: FnOnce() -> Result<R, &'static str> + UnwindSafe>(
&self,
initialize_block_fn: IB,
at: &BlockId<B>,
method: &str,
call_data: &[u8],
changes: &RefCell<OverlayedChanges>,
initialize_block: InitializeBlock<'a, B>,
execution_manager: ExecutionManager<EM>,
native_call: Option<NC>,
side_effects_handler: Option<&mut O>,
proof_recorder: &Option<Rc<RefCell<ProofRecorder<B>>>>
) -> Result<NativeOrEncoded<R>> where
ExecutionManager<EM>: Clone,

Execute a contextual call on top of state in a block of a given hash.

No changes are made. Before executing the method, passed header is installed as the current header of the execution context.

fn runtime_version(&self, id: &BlockId<B>) -> Result<RuntimeVersion, Error>

Extract RuntimeVersion of given block

No changes are made.

fn call_at_state<O: Externalities, S: Backend<H>, F: FnOnce(Result<NativeOrEncoded<R>, Self::Error>, Result<NativeOrEncoded<R>, Self::Error>) -> Result<NativeOrEncoded<R>, Self::Error>, R: Encode + Decode + PartialEq, NC: FnOnce() -> Result<R, &'static str> + UnwindSafe>(
&self,
state: &S,
overlay: &mut OverlayedChanges,
method: &str,
call_data: &[u8],
manager: ExecutionManager<F>,
native_call: Option<NC>,
side_effects_handler: Option<&mut O>
) -> Result<(NativeOrEncoded<R>, S::Transaction, Option<MemoryDB<H>>), Error>

Execute a call to a contract on top of given state.

No changes are made.

fn prove_at_trie_state<S: TrieBackendStorage<H>>(
&self,
trie_state: &TrieBackend<S, H>,
overlay: &mut OverlayedChanges,
method: &str,
call_data: &[u8]
) -> Result<(Vec<u8>, Vec<Vec<u8>>), Error>

Execute a call to a contract on top of given trie state, gathering execution proof.

No changes are made.

fn native_runtime_version(&self) -> Option<&NativeVersion>

Get runtime version if supported.

Loading content...
Provided methods

fn prove_at_state<S: Backend<H>>(
&self,
state: S,
overlay: &mut OverlayedChanges,
method: &str,
call_data: &[u8]
) -> Result<(Vec<u8>, Vec<Vec<u8>>), Error>

Execute a call to a contract on top of given state, gathering execution proof.

No changes are made.

Loading content...