Struct substrate_client::Client

pub struct Client<B, E, Block, RA> where
    Block: BlockT,  { /* fields omitted */ }

Substrate Client

Methods

impl<B, E, Block, RA> Client<B, E, Block, RA> where
B: Backend<Block, Blake2Hasher>,
E: CallExecutor<Block, Blake2Hasher>,
Block: BlockT<Hash = H256>,

pub fn new<S: BuildStorage>(
backend: Arc<B>,
executor: E,
build_genesis_storage: S,
execution_strategies: ExecutionStrategies
) -> Result<Self>

Creates new Substrate Client with given blockchain and code executor.

pub fn execution_strategies(&self) -> &ExecutionStrategies

Get a reference to the execution strategies.

pub fn state_at(&self, block: &BlockId<Block>) -> Result<B::State>

Get a reference to the state at a given block.

pub fn backend(&self) -> &Arc<B>

Expose backend reference. To be used in tests only

pub fn storage_keys(
&self,
id: &BlockId<Block>,
key_prefix: &StorageKey
) -> Result<Vec<StorageKey>>

Return storage entry keys in state in a block of given hash with given prefix.

pub fn storage(
&self,
id: &BlockId<Block>,
key: &StorageKey
) -> Result<Option<StorageData>>

Return single storage entry of contract under given address in state in a block of given hash.

pub fn code_at(&self, id: &BlockId<Block>) -> Result<Vec<u8>>

Get the code at a given block.

pub fn runtime_version_at(&self, id: &BlockId<Block>) -> Result<RuntimeVersion>

Get the RuntimeVersion at a given block.

pub fn executor(&self) -> &E

Get call executor reference.

pub fn read_proof(
&self,
id: &BlockId<Block>,
key: &[u8]
) -> Result<Vec<Vec<u8>>>

Reads storage value at a given block + key, returning read proof.

pub fn execution_proof(
&self,
id: &BlockId<Block>,
method: &str,
call_data: &[u8]
) -> Result<(Vec<u8>, Vec<Vec<u8>>)>

Execute a call to a contract on top of state in a block of given hash AND returning execution proof.

No changes are made.

pub fn header_proof(
&self,
id: &BlockId<Block>
) -> Result<(Block::Header, Vec<Vec<u8>>)>

Reads given header and generates CHT-based header proof.

pub fn block_hash(
&self,
block_number: <<Block as BlockT>::Header as HeaderT>::Number
) -> Result<Option<Block::Hash>>

Get block hash by number.

pub fn header_proof_with_cht_size(
&self,
id: &BlockId<Block>,
cht_size: u64
) -> Result<(Block::Header, Vec<Vec<u8>>)>

Reads given header and generates CHT-based header proof for CHT of given size.

pub fn max_key_changes_range(
&self,
first: NumberFor<Block>,
last: BlockId<Block>
) -> Result<Option<(NumberFor<Block>, BlockId<Block>)>>

Get longest range within [first; last] that is possible to use in key_changes and key_changes_proof calls. Range could be shortened from the beginning if some changes tries have been pruned. Returns Ok(None) if changes trues are not supported.

pub fn key_changes(
&self,
first: NumberFor<Block>,
last: BlockId<Block>,
key: &StorageKey
) -> Result<Vec<(NumberFor<Block>, u32)>>

Get pairs of (block, extrinsic) where key has been changed at given blocks range. Works only for runtimes that are supporting changes tries.

pub fn key_changes_proof(
&self,
first: Block::Hash,
last: Block::Hash,
min: Block::Hash,
max: Block::Hash,
key: &StorageKey
) -> Result<ChangesProof<Block::Header>>

Get proof for computation of (block, extrinsic) pairs where key has been changed at given blocks range. min is the hash of the first block, which changes trie root is known to the requester - when we're using changes tries from ascendants of this block, we should provide proofs for changes tries roots max is the hash of the last block known to the requester - we can't use changes tries from descendants of this block. Works only for runtimes that are supporting changes tries.

pub fn key_changes_proof_with_cht_size(
&self,
first: Block::Hash,
last: Block::Hash,
min: Block::Hash,
max: Block::Hash,
key: &StorageKey,
cht_size: u64
) -> Result<ChangesProof<Block::Header>>

Does the same work as key_changes_proof, but assumes that CHTs are of passed size.

pub fn new_block(&self) -> Result<BlockBuilder<Block, Self>> where
E: Clone + Send + Sync,
RA: Send + Sync,
Self: ProvideRuntimeApi,
Self::Api: BlockBuilderAPI<Block>,

Create a new block, built on the head of the chain.

pub fn new_block_at(
&self,
parent: &BlockId<Block>
) -> Result<BlockBuilder<Block, Self>> where
E: Clone + Send + Sync,
RA: Send + Sync,
Self: ProvideRuntimeApi,
Self::Api: BlockBuilderAPI<Block>,

Create a new block, built on top of parent.

pub fn lock_import_and_run<R, Err, F>(&self, f: F) -> Result<R, Err> where
F: FnOnce(&mut ClientImportOperation<Block, Blake2Hasher, B>) -> Result<R, Err>,
Err: From<Error>,

Lock the import lock, and run operations inside.

pub fn set_head(&self, id: BlockId<Block>) -> Result<()>

Set a block as best block.

pub fn apply_head(
&self,
operation: &mut ClientImportOperation<Block, Blake2Hasher, B>,
id: BlockId<Block>
) -> Result<()>

Set a block as best block, and apply it to an operation.

pub fn apply_block(
&self,
operation: &mut ClientImportOperation<Block, Blake2Hasher, B>,
import_block: ImportBlock<Block>,
new_cache: HashMap<CacheKeyId, Vec<u8>>
) -> Result<ImportResult> where
E: CallExecutor<Block, Blake2Hasher> + Send + Sync + Clone,

Apply a checked and validated block to an operation. If a justification is provided then finalized must be true.

pub fn apply_aux<'a, 'b: 'a, 'c: 'a, I: IntoIterator<Item = &'a (&'c [u8], &'c [u8])>, D: IntoIterator<Item = &'a &'b [u8]>>(
&self,
operation: &mut ClientImportOperation<Block, Blake2Hasher, B>,
insert: I,
delete: D
) -> Result<()>

Apply auxiliary data insertion into an operation.

pub fn apply_finality(
&self,
operation: &mut ClientImportOperation<Block, Blake2Hasher, B>,
id: BlockId<Block>,
justification: Option<Justification>,
notify: bool
) -> Result<()>

Mark all blocks up to given as finalized in operation. If a justification is provided it is stored with the given finalized block (any other finalized blocks are left unjustified).

pub fn finalize_block(
&self,
id: BlockId<Block>,
justification: Option<Justification>,
notify: bool
) -> Result<()>

Finalize a block. This will implicitly finalize all blocks up to it and fire finality notifications.

Pass a flag to indicate whether finality notifications should be propagated. This is usually tied to some synchronization state, where we don't send notifications while performing major synchronization work.

pub fn revert(&self, n: NumberFor<Block>) -> Result<NumberFor<Block>>

Attempts to revert the chain by n blocks. Returns the number of blocks that were successfully reverted.

pub fn info(&self) -> Result<ClientInfo<Block>>

Get blockchain info.

pub fn block_status(&self, id: &BlockId<Block>) -> Result<BlockStatus>

Get block status.

pub fn header(
&self,
id: &BlockId<Block>
) -> Result<Option<<Block as BlockT>::Header>>

Get block header by id.

pub fn body(
&self,
id: &BlockId<Block>
) -> Result<Option<Vec<<Block as BlockT>::Extrinsic>>>

Get block body by id.

pub fn justification(
&self,
id: &BlockId<Block>
) -> Result<Option<Justification>>

Get block justification set by id.

pub fn block(&self, id: &BlockId<Block>) -> Result<Option<SignedBlock<Block>>>

Get full block by id.

pub fn best_block_header(&self) -> Result<<Block as BlockT>::Header>

Get best block header.

pub fn best_containing(
&self,
target_hash: Block::Hash,
maybe_max_number: Option<NumberFor<Block>>
) -> Result<Option<Block::Hash>>

Get the most recent block hash of the best (longest) chains that contain block with the given target_hash.

The search space is always limited to blocks which are in the finalized chain or descendents of it.

If maybe_max_block_number is Some(max_block_number) the search is limited to block numbers <= max_block_number. in other words as if there were no blocks greater max_block_number. TODO : we want to move this implement to blockchain::Backend, see #1443 Returns Ok(None) if target_hash is not found in search space. TODO: document time complexity of this, see #1444

pub fn uncles(
&self,
target_hash: Block::Hash,
max_generation: NumberFor<Block>
) -> Result<Vec<Block::Hash>>

Gets the uncles of the block with target_hash going back max_generation ancestors.